+++ src/test/java/io/beanmapper/testmodel/construct_not_matching/FlatConstructSource2.java
@@ -7,9 +7,11 @@ public class FlatConstructSource2 {
 
     public Long id;
     public String street;
+
     @BeanAlias("city")
     @BeanProperty(name = "nestedClass.city")
     public String city;
+
     @BeanAlias("country")
     @BeanProperty(name = "nestedClass.country")
     public String country;